requirements.txt
langchain
langchain-community
langchain-openai
streamlit
streamlit_chat
python-dotenv